Изменение double на long  расчетах суммы. Исправление проверки завешения работы тасков экзекьютора и их закрытие. Изменение моделей. Добавление класса тестирования для разоного рода подключенией. Перекладка расчета саблистов по пулу кредитов, а не по одному за раз. Исправление классов работы с БД отправителем и классов по работе с внутренними массивами. Исзменние алгоритмов расчетов задолженностей в основном классе CalcAllDebts

+++ src/main/java/ru/intel/credits/calc/CalcDebtsStrategy.java
@@ -6,5 +6,5 @@ import ru.intel.credits.model.VidOperDog;
 import java.util.HashMap;
 
 public interface CalcDebtsStrategy {
-    public double calcDebt(PrCred cred, HashMap<Long, VidOperDog> opers, long idDebt);
+    public long calcDebt(PrCred cred, HashMap<Long, VidOperDog> opers, long idDebt);
 }
